# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API",
        "description": "Descripción detallada de la API",
        "contact": {
            "email": "francisco.gutierrez.g@live.cl"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/fastpay-app/public/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/payments/merchant/{id}": {
            "get": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Obtener pagos por ID de comercio",
                "description": "Devuelve los pagos asociados a un comercio específico por su ID.",
                "operationId": "showPaymentsByMerchantId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del comercio",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagos encontrados para el comercio especificado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pagos encontrados para comercio ID 3"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 121
                                                    },
                                                    "merchant_id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "amount": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 6956
                                                    },
                                                    "fee": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 41.74
                                                    },
                                                    "payment_method": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "approved"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-12-10T01:32:29.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-12-10T01:32:29.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron pagos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontraron pagos asociados al comercio ID X."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Acceso denegado al recurso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/payments/merchant/{merchantId}/payment/{paymentId}": {
            "get": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Obtener un pago por ID de comercio y ID de pago",
                "description": "Devuelve un pago específico asociado a un comercio dado por su ID y el ID de pago.",
                "operationId": "showPaymentByMerchantIdAndPaymentId",
                "parameters": [
                    {
                        "name": "merchantId",
                        "in": "path",
                        "description": "ID del comercio",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    },
                    {
                        "name": "paymentId",
                        "in": "path",
                        "description": "ID del pago",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 121
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pago encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pago encontrado para comercio ID X."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 121
                                                },
                                                "merchant_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 6956
                                                },
                                                "fee": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 41.74
                                                },
                                                "payment_method": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "approved"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-10T01:32:29.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-10T01:32:29.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Acceso denegado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontró un pago coincidente con la busqueda.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontró un pago coincidente con la busqueda."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}